requirements.txt
matplotlib==3.8.2
opencv_contrib_python==4.8.0.76
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
psycopg2_binary==2.9.9
python_bcrypt==0.3.2
PyYAML==6.0.1
PyYAML==6.0.1
streamlit==1.32.2
streamlit_authenticator==0.3.2
tensorflow==2.16.1

